pyproject.toml
[project]
name = "choice-paralysis"
version = "0.1.0"
description = "Backlog Gladiator"
authors = [
    {name = "Alex Yanchenko", email = "oleksandr.yanchenko.ca@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "greenlet>=3.2.4",
    "punq>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["choice_paralysis"]

[dependency-groups]
dev = [
    "ruff (>=0.14.5,<0.15.0)",
    "ty>=0.0.1a27",
    "pre-commit>=4.0.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[tool.ruff]
line-length = 110
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D203", "D212"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
